Fix: 질문자, 트레이더 문의 상세 페이지 이전글 다음글 오류 수정, 예외처리 수정 #199

+++ src/main/java/com/be3c/sysmetic/domain/member/service/InquiryServiceImpl.java
@@ -18,7 +18,12 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+
+import static com.be3c.sysmetic.domain.member.message.NoticeDeleteFailMessage.NOT_FOUND_INQUIRY;
+import static com.be3c.sysmetic.domain.member.message.NoticeDeleteFailMessage.NOT_FOUND_NOTICE;
 
 @Service
 @Transactional(readOnly = true)
@@ -39,7 +44,7 @@ public class InquiryServiceImpl implements InquiryService {
     // 문의 단건 조회
     @Override
     public Inquiry findOneInquiry(Long inquiryId) {
-        return inquiryRepository.findById(inquiryId).orElseThrow(EntityNotFoundException::new);
+        return inquiryRepository.findById(inquiryId).orElseThrow(() -> new EntityNotFoundException("문의가 없습니다."));
     }
 
     // 문의 전체 조회
@@ -104,7 +109,7 @@ public class InquiryServiceImpl implements InquiryService {
     @Transactional
     public boolean modifyInquiry(Long inquiryId, String inquiryTitle, String inquiryContent) {
 
-        Inquiry inquiry = inquiryRepository.findById(inquiryId).orElseThrow(EntityNotFoundException::new);
+        Inquiry inquiry = inquiryRepository.findById(inquiryId).orElseThrow(() -> new EntityNotFoundException("문의가 없습니다."));
         if (inquiry.getInquiryStatus() == InquiryStatus.unclosed) {
             inquiry.setInquiryTitle(inquiryTitle);
             inquiry.setInquiryContent(inquiryContent);
@@ -121,7 +126,7 @@ public class InquiryServiceImpl implements InquiryService {
     @Transactional
     public boolean deleteInquiry(Long inquiryId) {
 
-        Inquiry inquiry = inquiryRepository.findById(inquiryId).orElseThrow(EntityNotFoundException::new);
+        Inquiry inquiry = inquiryRepository.findById(inquiryId).orElseThrow(() -> new EntityNotFoundException("문의가 없습니다."));
 
         if (inquiry.getInquiryStatus() == InquiryStatus.unclosed) {
             inquiryRepository.delete(inquiry);
@@ -137,7 +142,7 @@ public class InquiryServiceImpl implements InquiryService {
     @Transactional
     public boolean deleteAdminInquiry(Long inquiryId) {
 
-        Inquiry inquiry = inquiryRepository.findById(inquiryId).orElseThrow(EntityNotFoundException::new);
+        Inquiry inquiry = inquiryRepository.findById(inquiryId).orElseThrow(() -> new EntityNotFoundException("문의가 없습니다."));
 
         inquiryRepository.delete(inquiry);
 
@@ -148,16 +153,57 @@ public class InquiryServiceImpl implements InquiryService {
     // 관리자 목록 삭제
     @Override
     @Transactional
-    public Integer deleteAdminInquiryList(List<Long> inquiryIdList) {
+    public Map<Long, String> deleteAdminInquiryList(List<Long> inquiryIdList) {
 
-        return inquiryRepository.bulkDelete(inquiryIdList);
+        if (inquiryIdList == null || inquiryIdList.isEmpty()) {
+            throw new EntityNotFoundException("문의가 한 개도 선택되지 않았습니다.");
+        }
+
+        Map<Long, String> failDelete = new HashMap<>();
+
+        for (Long inquiryId : inquiryIdList) {
+            try {
+                inquiryRepository.findById(inquiryId);
+            }
+            catch (EntityNotFoundException e) {
+                failDelete.put(inquiryId, NOT_FOUND_INQUIRY.getMessage());
+            }
+        }
+
+        inquiryRepository.bulkDelete(inquiryIdList);
+
+        return failDelete;
     }
 
 
-    // 이전 문의 제목 조회
+    // 트레이더별 문의id로 조회
     @Override
-    public String findPreviousInquiryTitle(Long inquiryId) {
-        List<Inquiry> previousInquiryList = inquiryRepository.findPreviousInquiry(inquiryId, PageRequest.of(0, 1));
+    public Inquiry findInquiryByTraderIdAndInquiryId(Long inquiryId, Long traderId) {
+        List<Inquiry> inquiryList = inquiryRepository.findInquiryByTraderIdAndInquiryId(inquiryId, traderId, PageRequest.of(0, 1));
+
+        if (inquiryList.isEmpty()) {
+            throw new EntityNotFoundException("문의를 찾을 수 없습니다.");
+        } else {
+            return inquiryList.get(0);
+        }
+    }
+
+    // 질문자별 문의id로 조회
+    @Override
+    public Inquiry findInquiryByInquirerIdAndInquiryId(Long inquiryId, Long inquirerId) {
+        List<Inquiry> inquiryList = inquiryRepository.findInquiryByInquirerIdAndInquiryId(inquiryId, inquirerId, PageRequest.of(0, 1));
+
+        if (inquiryList.isEmpty()) {
+            throw new EntityNotFoundException("문의를 찾을 수 없습니다.");
+        } else {
+            return inquiryList.get(0);
+        }
+    }
+
+    // 관리자 이전 문의 제목 조회
+    @Override
+    public String adminFindPreviousInquiryTitle(Long inquiryId) {
+        List<Inquiry> previousInquiryList = inquiryRepository.adminFindPreviousInquiry(inquiryId, PageRequest.of(0, 1));
 
         String previousInquiryTitle;
         if (previousInquiryList.isEmpty()) {
@@ -170,10 +216,10 @@ public class InquiryServiceImpl implements InquiryService {
         return previousInquiryTitle;
     }
 
-    // 이전 문의 작성일 조회
+    // 관리자 이전 문의 작성일 조회
     @Override
-    public LocalDateTime findPreviousInquiryWriteDate(Long inquiryId) {
-        List<Inquiry> previousInquiryList = inquiryRepository.findPreviousInquiry(inquiryId, PageRequest.of(0, 1));
+    public LocalDateTime adminFindPreviousInquiryWriteDate(Long inquiryId) {
+        List<Inquiry> previousInquiryList = inquiryRepository.adminFindPreviousInquiry(inquiryId, PageRequest.of(0, 1));
 
         LocalDateTime previousInquiryWriteDate;
         if (previousInquiryList.isEmpty()) {
@@ -186,10 +232,10 @@ public class InquiryServiceImpl implements InquiryService {
         return previousInquiryWriteDate;
     }
 
-    // 다음 문의 제목 조회
+    // 관리자 다음 문의 제목 조회
     @Override
-    public String findNextInquiryTitle(Long inquiryId) {
-        List<Inquiry> nextInquiryList = inquiryRepository.findNextInquiry(inquiryId, PageRequest.of(0, 1));
+    public String adminFindNextInquiryTitle(Long inquiryId) {
+        List<Inquiry> nextInquiryList = inquiryRepository.adminFindNextInquiry(inquiryId, PageRequest.of(0, 1));
 
         String nextInquiryTitle;
         if (nextInquiryList.isEmpty()) {
@@ -202,10 +248,138 @@ public class InquiryServiceImpl implements InquiryService {
         return nextInquiryTitle;
     }
 
-    // 다음 문의 제목 조회
+    // 관리자 다음 문의 제목 조회
     @Override
-    public LocalDateTime findNextInquiryWriteDate(Long inquiryId) {
-        List<Inquiry> nextInquiryList = inquiryRepository.findNextInquiry(inquiryId, PageRequest.of(0, 1));
+    public LocalDateTime adminFindNextInquiryWriteDate(Long inquiryId) {
+        List<Inquiry> nextInquiryList = inquiryRepository.adminFindNextInquiry(inquiryId, PageRequest.of(0, 1));
+
+        LocalDateTime nextInquiryWriteDate;
+        if (nextInquiryList.isEmpty()) {
+            nextInquiryWriteDate = null;
+        } else {
+            Inquiry previousInquiry = nextInquiryList.get(0);
+            nextInquiryWriteDate = previousInquiry.getInquiryRegistrationDate();
+        }
+
+        return nextInquiryWriteDate;
+    }
+
+    // 트레이더 이전 문의 제목 조회
+    @Override
+    public String traderFindPreviousInquiryTitle(Long inquiryId, Long traderId) {
+        List<Inquiry> previousInquiryList = inquiryRepository.traderFindPreviousInquiry(inquiryId, traderId, PageRequest.of(0, 1));
+
+        String previousInquiryTitle;
+        if (previousInquiryList.isEmpty()) {
+            previousInquiryTitle = null;
+        } else {
+            Inquiry previousInquiry = previousInquiryList.get(0);
+            previousInquiryTitle = previousInquiry.getInquiryTitle();
+        }
+
+        return previousInquiryTitle;
+    }
+
+    // 트레이더 이전 문의 작성일 조회
+    @Override
+    public LocalDateTime traderFindPreviousInquiryWriteDate(Long inquiryId, Long traderId) {
+        List<Inquiry> previousInquiryList = inquiryRepository.traderFindPreviousInquiry(inquiryId, traderId, PageRequest.of(0, 1));
+
+        LocalDateTime previousInquiryWriteDate;
+        if (previousInquiryList.isEmpty()) {
+            previousInquiryWriteDate = null;
+        } else {
+            Inquiry previousInquiry = previousInquiryList.get(0);
+            previousInquiryWriteDate = previousInquiry.getInquiryRegistrationDate();
+        }
+
+        return previousInquiryWriteDate;
+    }
+
+    // 트레이더 다음 문의 제목 조회
+    @Override
+    public String traderFindNextInquiryTitle(Long inquiryId, Long traderId) {
+        List<Inquiry> nextInquiryList = inquiryRepository.traderFindNextInquiry(inquiryId, traderId, PageRequest.of(0, 1));
+
+        String nextInquiryTitle;
+        if (nextInquiryList.isEmpty()) {
+            nextInquiryTitle = null;
+        } else {
+            Inquiry previousInquiry = nextInquiryList.get(0);
+            nextInquiryTitle = previousInquiry.getInquiryTitle();
+        }
+
+        return nextInquiryTitle;
+    }
+
+    // 트레이더 다음 문의 제목 조회
+    @Override
+    public LocalDateTime traderFindNextInquiryWriteDate(Long inquiryId, Long traderId) {
+        List<Inquiry> nextInquiryList = inquiryRepository.traderFindNextInquiry(inquiryId, traderId, PageRequest.of(0, 1));
+
+        LocalDateTime nextInquiryWriteDate;
+        if (nextInquiryList.isEmpty()) {
+            nextInquiryWriteDate = null;
+        } else {
+            Inquiry previousInquiry = nextInquiryList.get(0);
+            nextInquiryWriteDate = previousInquiry.getInquiryRegistrationDate();
+        }
+
+        return nextInquiryWriteDate;
+    }
+
+    // 질문자 이전 문의 제목 조회
+    @Override
+    public String inquirerFindPreviousInquiryTitle(Long inquiryId, Long inquirerId) {
+        List<Inquiry> previousInquiryList = inquiryRepository.inquirerFindPreviousInquiry(inquiryId, inquirerId, PageRequest.of(0, 1));
+
+        String previousInquiryTitle;
+        if (previousInquiryList.isEmpty()) {
+            previousInquiryTitle = null;
+        } else {
+            Inquiry previousInquiry = previousInquiryList.get(0);
+            previousInquiryTitle = previousInquiry.getInquiryTitle();
+        }
+
+        return previousInquiryTitle;
+    }
+
+    // 질문자 이전 문의 작성일 조회
+    @Override
+    public LocalDateTime inquirerFindPreviousInquiryWriteDate(Long inquiryId, Long inquirerId) {
+        List<Inquiry> previousInquiryList = inquiryRepository.inquirerFindPreviousInquiry(inquiryId, inquirerId, PageRequest.of(0, 1));
+
+        LocalDateTime previousInquiryWriteDate;
+        if (previousInquiryList.isEmpty()) {
+            previousInquiryWriteDate = null;
+        } else {
+            Inquiry previousInquiry = previousInquiryList.get(0);
+            previousInquiryWriteDate = previousInquiry.getInquiryRegistrationDate();
+        }
+
+        return previousInquiryWriteDate;
+    }
+
+    // 질문자 다음 문의 제목 조회
+    @Override
+    public String inquirerFindNextInquiryTitle(Long inquiryId, Long inquirerId) {
+        List<Inquiry> nextInquiryList = inquiryRepository.inquirerFindNextInquiry(inquiryId, inquirerId, PageRequest.of(0, 1));
+
+        String nextInquiryTitle;
+        if (nextInquiryList.isEmpty()) {
+            nextInquiryTitle = null;
+        } else {
+            Inquiry previousInquiry = nextInquiryList.get(0);
+            nextInquiryTitle = previousInquiry.getInquiryTitle();
+        }
+
+        return nextInquiryTitle;
+    }
+
+    // 질문자 다음 문의 제목 조회
+    @Override
+    public LocalDateTime inquirerFindNextInquiryWriteDate(Long inquiryId, Long inquirerId) {
+        List<Inquiry> nextInquiryList = inquiryRepository.inquirerFindNextInquiry(inquiryId, inquirerId, PageRequest.of(0, 1));
 
         LocalDateTime nextInquiryWriteDate;
         if (nextInquiryList.isEmpty()) {
@@ -222,7 +396,7 @@ public class InquiryServiceImpl implements InquiryService {
     // 관리자 검색 조회
     // 전체, 답변 대기, 답변 완료
     // 검색 (전략명, 트레이더, 질문자)
-    public Page<Inquiry> findInquiresAdmin(InquiryAdminListShowRequestDto inquiryAdminListShowRequestDto, Integer page) {
+    public Page<Inquiry> findInquiriesAdmin(InquiryAdminListShowRequestDto inquiryAdminListShowRequestDto, Integer page) {
 
         return inquiryRepository.adminInquirySearchWithBooleanBuilder(inquiryAdminListShowRequestDto, PageRequest.of(page, 10));
     }
@@ -231,7 +405,7 @@ public class InquiryServiceImpl implements InquiryService {
     // 문의자, 트레이더 검색 조회
     // 정렬 순 셀렉트 박스 (최신순, 전략명)
     // 답변상태 셀렉트 박스 (전체, 답변 대기, 답변 완료)
-    public Page<Inquiry> findInquires(InquiryListShowRequestDto inquiryListShowRequestDto, Integer page) {
+    public Page<Inquiry> findInquiries(InquiryListShowRequestDto inquiryListShowRequestDto, Integer page) {
 
         return inquiryRepository.inquirySearchWithBooleanBuilder(inquiryListShowRequestDto, PageRequest.of(page, 10));
     }
